feat: jwt user authentication

+++ api/controllers/users.controller.js
@@ -1,6 +1,7 @@
 const User = require("../models/user.model");
 const mailer = require("../config/mailer.config");
 const createError = require("http-errors");
+const jwt = require("jsonwebtoken");
 
 module.exports.list = (req, res, next) => {
   User.find()
@@ -37,6 +38,10 @@ module.exports.confirm = (req, res, next) => {
 };
 
 module.exports.update = (req, res, next) => {
+  if (req.user.id !== req.params.id) {
+    return next(createError(403, "Forbidden"));
+  }
+  
   Object.assign(req.user, req.body);
   req.user.save()
     .then((user) => res.json(user))
@@ -44,6 +49,10 @@ module.exports.update = (req, res, next) => {
 };
 
 module.exports.delete = (req, res, next) => {
+  if (req.user.id !== req.params.id) {
+    return next(createError(403, "Forbidden"));
+  }
+
   User.deleteOne({ _id: req.user.id })
     .then(() => res.status(204).send())
     .catch(next);
@@ -52,8 +61,12 @@ module.exports.delete = (req, res, next) => {
 module.exports.login = (req, res, next) => {
   User.findOne({ email: req.body.email })
     .then((user) => {
-      if (!user) {
+      if (!user || !req.body.password) {
         return next(createError(401, "Invalid credentials"));
+      }
+
+      if (!user.confirm) {
+        return next(creterError(401, "Please confirm your account"));
       }
 
       user.checkPassword(req.body.password)
@@ -61,6 +74,8 @@ module.exports.login = (req, res, next) => {
           if (!match) {
             return next(createError(401, "Invalid credentials"));
           }
+
+          const token = jwt.sign({ sub: user.id, exp: Date.now() / 1000 + 3600 }, process.env.JWT_SECRET);
 
           res.json(user);
         });
